Broke out sidebar into several components
