+++ client/src/components/sidebar-components/SidebarDMList.tsx
@@ -1,0 +1,5 @@
+export default function SidebarDMList() {
+  return (
+    <div>SidebarDMList</div>
+  )
+}
