+++ client/src/components/sidebar-components/SidebarUserList.tsx
@@ -1,0 +1,5 @@
+export default function SidebarUserList() {
+  return (
+    <div>SidebarUserList</div>
+  )
+}
